Multiple prefixes.

+++ src/commands/hug.ts
@@ -8,12 +8,7 @@
 .##.......##..........##........####.##....##..######..
 */
 export const command: Command = {
-    run: async (client, msg) => {
-        if (!msg.mentions.members.first()) return msg.reply('you need to tag someone to hug 🤗');
-
-        await msg.delete();
-        return msg.channel.send(`${msg.author} hugged ${msg.mentions.members.first()} 🤗💞`)
-    },
+    run: async (_client, msg) => !msg.mentions.members.first() ? msg.reply('you need to tag someone to hug 🤗') : await msg.delete() && msg.channel.send(`${msg.author} hugged ${msg.mentions.members.first()} 🤗💞`),
     config: {
         name: 'hug',
         args: true,
